+++ packages/wiseclient/lib/src/secure_token_storage/fresh_secure_token_storage.dart
@@ -1,8 +1,8 @@
 import 'dart:convert' show jsonEncode;
 
 import 'package:flutter_secure_storage/flutter_secure_storage.dart';
+import 'package:fresh_dio/fresh_dio.dart';
 import 'package:wiseclient/src/token_model/oauth_token.dart';
-import '../fresh/fresh.dart';
 import 'map_from_token_extension.dart';
 import 'token_from_string_extension.dart';
 
@@ -12,10 +12,15 @@ class FreshSecureTokenStorage implements TokenStorage<OAuthToken> {
   OAuthToken? _token;
 
   /// [FlutterSecureStorage] to save token on device
-  final storage = const FlutterSecureStorage();
+  final storage = const FlutterSecureStorage(
+    iOptions: IOSOptions(
+      accessibility: KeychainAccessibility.first_unlock,
+      synchronizable: true,
+    ),
+  );
 
   /// String to save to [FlutterSecureStorage]
-  final storageIdentifier = 'OAUTH_TOKEN';
+  final storageIdentifier = 'OAUTH_TOKEN_STORAGE';
 
   @override
   Future<void> delete() async {